Modularize AUR commands by separating them into individual files and reorganizing the namespace structure under `Commands.Aur`. Update `Program.cs` and reflection directives accordingly.
